Each module compiles on its own TODO: lots of OS_* for Pin CRT

+++ src/PINdemonium/Config.cpp
@@ -98,9 +98,7 @@ string Config::getHeapDir(){
 	return this->heap_dir;
 }
 
-
-
-long double Config::getDumpNumber(){
+int Config::getDumpNumber(){
 	return this->dump_number;
 }
